requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pyspark==3.5.6
apache-airflow-providers-common-compat
apache-airflow-providers-apache-spark
requests>=2.28.0
pyarrow>=10.0.0
pandas>=1.5.0